+++ src/Annotation.h
@@ -70,6 +70,7 @@ struct AnnotCreateArgs {
     AnnotationType annotType = AnnotationType::Unknown;
     // the following are set depending on type of the annotation
     ParsedColor col;
+    bool copyToClipboard = false;
 };
 
 int PageNo(Annotation*);